pdf-extractor-analyzer/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-extractor-analyzer"
version = "0.1.0"
description = "Image-first PDF extraction and analysis pipeline with Replicate vision models"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pymupdf>=1.24.0",
  "pillow>=10.0.0",
  "replicate>=0.25.0",
  "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
pdf-extractor = "pdf_extractor_analyzer.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
